fix: tests for refactored email submission code
