+++ src/app/modules/submission/ParsedResponsesObject.class.ts
@@ -18,7 +18,6 @@ import {
   createSingpassParsedResponses,
 } from '../spcp/spcp.util'
 
-import { newEncryptionBoundaryFlag } from './encrypt-submission/encrypt-submission.constants'
 import {
   ConflictError,
   ProcessingError,
@@ -151,7 +150,7 @@ export default class ParsedResponsesObject {
           // TODO(#780): Remove this once submission service is separated into
           // Email and Encrypted services
           form.responseMode === FormResponseMode.Encrypt &&
-          !form.get(newEncryptionBoundaryFlag)
+          !form.newEncryptionBoundary
             ? 'answer' in response &&
               typeof response.answer === 'string' &&
               response.answer.trim() !== ''
